Adding the Xunit testing for the program
implement the following tests CheckRegister, CheckLogin,
CheckFlightBooking, CancelFlightBooking, DisplayUserBooking, and ModifyUserBooking
using InlineData, autofixture, and Fluent Assertions
and i haved refactored some methods in the BookingManager class
(ModifyBooking) to improve the code with clean code and FP principles
