+++ AirportTicketBookingExercise/AirportTicketBookingExercise.csproj
@@ -7,4 +7,9 @@
         <Nullable>enable</Nullable>
     </PropertyGroup>
 
+    <ItemGroup>
+        <PackageReference Include="AutoFixture" Version="5.0.0-preview0011"/>
+        <PackageReference Include="Moq" Version="4.20.72"/>
+    </ItemGroup>
+
 </Project>
